+++ src/main/java/com/aaronpmaus/jProt/protein/Protein.java
@@ -1,7 +1,11 @@
 package com.aaronpmaus.jProt.protein;
 
 import com.aaronpmaus.jProt.io.*;
+import com.aaronpmaus.jProt.sequence.*;
+
 import com.aaronpmaus.jMath.graph.*;
+import com.aaronpmaus.jMath.transformations.Transformable;
+import com.aaronpmaus.jMath.transformations.Transformation;
 
 import java.util.HashMap;
 import java.util.ArrayList;
@@ -34,7 +38,7 @@ import java.io.InputStream;
 * @version 0.6.0
 * @since 0.6.0
 */
-public class Protein implements Iterable<PolypeptideChain>{
+public class Protein implements Iterable<PolypeptideChain>, Transformable{
   private ArrayList<PolypeptideChain> chains;
   private String proteinName;
   private ArrayList<Bond> disulfideBonds;
@@ -205,12 +209,12 @@ public class Protein implements Iterable<PolypeptideChain>{
   *
   * @return a String containing the single letter residue names with no spaces
   */
-  public String getSequence(){
+  public ProteinSequence getSequence(){
     String seq = "";
     for(PolypeptideChain chain : this.chains){
-      seq += chain.getSequence();
+      seq += chain.getSequence().getSequenceString();
     }
-    return seq;
+    return new ProteinSequence(seq);
   }
 
   /**
@@ -329,6 +333,13 @@ public class Protein implements Iterable<PolypeptideChain>{
   */
   public boolean hydrogensEnabled(){
     return this.hydrogensEnabled;
+  }
+
+  @Override
+  public void applyTransformation(Transformation t){
+    for(PolypeptideChain chain : this ){
+      chain.applyTransformation(t);
+    }
   }
 
   @Override